# docs/integrations/openapi-gateway.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DIAL Gateway — integration surface (skeleton)",
    "version": "0.1.0",
    "description": "S134 skeleton for API-key drop-in readiness. Documents health + inbound webhook shapes. Never includes secret values. Health JSON note SoR: buildIntegrationsHealthNote (info.x-dial-sor.healthNote); docs pointer INTEGRATIONS_NOTE_BUILDER_SOR_DOCS (info.x-dial-sor.noteBuilderDocs). Webhooks: signature verify + claimProcessedEvent idempotency (info.x-dial-sor.webhookSignature / webhookIdempotency). Authoritative eng behaviour remains Pack/adapters; this is a contract outline for vendors/ops.",
    "x-dial-sor": {
      "probes": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATION_PROBE_KEYS",
      "envGroups": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATION_ENV_GROUPS",
      "envGroupLabels": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATION_ENV_GROUP_LABELS",
      "healthNote": "apps/gateway-web/src/lib/integrationsReadiness.ts#buildIntegrationsHealthNote",
      "healthNoteUiMax": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATIONS_HEALTH_NOTE_UI_MAX",
      "noteBuilderHint": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATIONS_NOTE_BUILDER_SOR_HINT_ID",
      "noteBuilderDocs": "apps/gateway-web/src/lib/integrationsReadiness.ts#INTEGRATIONS_NOTE_BUILDER_SOR_DOCS",
...
      "docs": "docs/integrations/README.md",
...
    }
  },
  "servers": [
    {
...
    }
  ],
...
  "paths": {
...
    "/api/webhooks/paynow": {
      "post": {
        "tags": ["webhooks"],
        "summary": "Paynow SHA512 result notification",
        "operationId": "webhookPaynow",
        "parameters": [
          {
            "name": "Hash",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "description": "Paynow hash (also validated from body fields)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookOpaqueBody" }
            },
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/WebhookOpaqueBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted or duplicate idempotent (signature verified)"
          },
          "401": { "description": "Bad signature" },
          "503": { "description": "Fail-closed (sandbox/live misconfig)" }
        }
      }
    },
...
    "/api/webhooks/ecocash": {
      "post": {
        "tags": ["webhooks"],
        "summary": "EcoCash HMAC webhook",
        "operationId": "webhookEcocash",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WebhookOpaqueBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted or duplicate idempotent (signature verified)"
          },
          "401": { "description": "Bad signature" },
          "503": { "description": "Fail-closed (sandbox/live misconfig)" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "InternalApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "INTERNAL_API_SECRET — never expose in browser bundles"
      }
    },
    "schemas": {
...
      "WebhookOpaqueBody": {
        "type": "object",
        "description": "Vendor payload — verified by adapter signature before mutate; idempotent via processed_events (S196 SoR: info.x-dial-sor.webhookSignature + webhookIdempotency)",
        "additionalProperties": true
      }
    }
  }
}